refactor(ui-react): extract office and terminal previews
